lat,lngの型をdoubleからfloatへ変更

+++ db/migrate/20170804062510_add_lng_to_advices.rb
@@ -1,5 +1,5 @@
 class AddLngToAdvices < ActiveRecord::Migration[5.1]
   def change
-    add_column :advices, :lng, :double
+    add_column :advices, :lng, :float
   end
 end
